Update browser extension UI with settings subpages (#1144)
