+++ apps/browser-extension/src/i18n/locales/en.json
@@ -340,6 +340,15 @@
     "autoLock8Hours": "8 hours",
     "autoLock24Hours": "24 hours",
     "versionPrefix": "Version ",
+    "preferences": "Preferences",
+    "autofillSettings": "Autofill Settings",
+    "clipboardSettings": "Clipboard Settings",
+    "contextMenuSettings": "Context Menu Settings",
+    "contextMenu": "Context Menu",
+    "contextMenuEnabled": "Context menu is enabled",
+    "contextMenuDisabled": "Context menu is disabled",
+    "contextMenuDescription": "Right-click on input fields to access AliasVault options",
+    "selectLanguage": "Select Language",
     "validation": {
       "apiUrlRequired": "API URL is required",
       "apiUrlInvalid": "Please enter a valid API URL",
